Use a common class Wrapper and IterWrapper to wrapp native resource.
Native resource (pointer to shared_ptr) is stored as a long in the
`Wrapper.Resource` class.

As Wrapper implements AutoCloseable and we register the (java) resource
to be cleaned at object destruction, we now properly delete the native
resource.


This also adding new macro to avoid writting the class name and so
reduce potential typos.

# Conflicts:
#	lib/build.gradle
#	lib/src/main/java/org/kiwix/libkiwix/Library.java